infra/database/scripts/requirements.txt
boto3==1.34.131
pandas==2.2.3
pyarrow==16.1.0
psycopg2-binary==2.9.9
requests==2.31.0
